refactoring

+++ src/main/java/com/example/operation/impl/TransferAccountCommand.java
@@ -20,9 +20,9 @@ public class TransferAccountCommand implements OperationCommand {
     public void execute() {
         try {
             System.out.println("Enter source account id:");
-            int fromAccountId = Integer.parseInt(scanner.nextLine());
+            Long fromAccountId = Long.parseLong(scanner.nextLine());
             System.out.println("Enter destination account id:");
-            int toAccountId = Integer.parseInt(scanner.nextLine());
+            Long toAccountId = Long.parseLong(scanner.nextLine());
             System.out.println("Enter amount to transfer:");
             BigDecimal amountToTransfer = BigDecimal.valueOf(Long.parseLong(scanner.nextLine()));
             accountService.transfer(fromAccountId, toAccountId, amountToTransfer);
